finished lab03

+++ lab/lab03/src/adventure/SpeciesListStage.java
@@ -92,7 +92,7 @@ public class SpeciesListStage implements AdventureStage {
      *
      * Similarity is defined as when listTwo contains all the elements in listOne.
      */
-    public static int arraySimilarity(List<String> listOne, List<String> listTwo) {
+    public static double arraySimilarity(List<String> listOne, List<String> listTwo) {
         List<String> copy = new ArrayList<>(listOne);
         int similarObjects = 0;
         for (String o : listTwo) {
@@ -101,6 +101,6 @@ public class SpeciesListStage implements AdventureStage {
                 copy.remove(o);
             }
         }
-        return similarObjects / listOne.size();
+        return (double) similarObjects / listOne.size();
     }
 }
